Move testdir to legacypath plugin
