+++ doc/en/conf.py
@@ -446,3 +446,7 @@ def setup(app: "sphinx.application.Sphinx") -> None:
         )
 
     sphinx.pycode.parser.VariableCommentPicker.is_final = patched_is_final
+
+    # legacypath.py monkey-patches pytest.Testdir in. Import the file so
+    # that autodoc can discover references to it.
+    import _pytest.legacypath  # noqa: F401
