+++ src/pytest/__init__.py
@@ -46,7 +46,6 @@ from _pytest.pytester import LineMatcher
 from _pytest.pytester import Pytester
 from _pytest.pytester import RecordedHookCall
 from _pytest.pytester import RunResult
-from _pytest.pytester import Testdir
 from _pytest.python import Class
 from _pytest.python import Function
 from _pytest.python import Instance
@@ -145,7 +144,6 @@ __all__ = [
     "StashKey",
     "version_tuple",
     "TempPathFactory",
-    "Testdir",
     "TempdirFactory",
     "UsageError",
     "WarningsRecorder",
